refactor(dispatcher): named versions for configs and classes fixes #2
using `instant/mixedpickupdropoff` for `v1` and `instant/sequential` for `v2`
