+++ prototype/dispatch/order-dispatcher/package.json
@@ -8,9 +8,9 @@
     "doc": "docs"
   },
   "scripts": {
-    "generate-deployment": "rm -rf build && mkdir -p build/config && ./mvnw package && cp target/dispatcher-runner.jar build/ && cp deploy/application.properties build/config/ && cp src/main/resources/dispatchSolverConfig-v1.xml build/config/",
+    "generate-deployment": "rm -rf build && mkdir -p build/config/instant/sequential && ./mvnw package && cp target/dispatcher-runner.jar build/ && cp deploy/application.properties build/config/ && cp src/main/resources/instant/sequentialdispatchSolverConfig.xml build/config/instant/sequential",
     "build": "yarn generate-deployment && cd build && zip -FSrq dispatcher.zip *",
-    "build:ec2": "sudo rm -rf build && mkdir -p build/config && ./mvnw package && cp target/dispatcher-runner.jar build/ && cp src/main/resources/application.properties build/config/ && cp src/main/resources/dispatchSolverConfig-v1.xml build/",
+    "build:ec2": "sudo rm -rf build && mkdir -p build/config && ./mvnw package && cp target/dispatcher-runner.jar build/ && cp src/main/resources/application.properties build/config/ && cp src/main/resources/instant/sequential/dispatchSolverConfig.xml build/",
     "start": "cd build && sudo java -jar dispatcher-runner.jar",
     "build:docker": "./build-docker.sh",
     "run:docker": "docker run -d --name dispatcher --rm -p 80:9999 order-dispatcher",
