+++ prototype/dispatch/order-dispatcher/src/main/java/com/aws/proto/dispatching/domain/planningentity/instant/mixedpickupdropoff/VisitOrDriver.java
@@ -10,10 +10,10 @@
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,7 +23,7 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.aws.proto.dispatching.domain.planningentity.v1;
+package com.aws.proto.dispatching.domain.planningentity.instant.mixedpickupdropoff;
 
 import com.aws.proto.dispatching.util.Constants;
 import org.optaplanner.core.api.domain.entity.PlanningEntity;
@@ -34,12 +34,14 @@ public interface VisitOrDriver {
 
     /**
      * Get the driver assigned to this planning entity.
+     *
      * @return The PlanningDriver. Sometimes {@code null}.
      */
     PlanningDriver getPlanningDriver();
 
     /**
      * Get the next visit assigned to this planning entity
+     *
      * @return The next PlanningVisit. Sometimes {@code null}.
      */
     @InverseRelationShadowVariable(sourceVariableName = Constants.PreviousVisitOrDriver)
@@ -47,6 +49,7 @@ public interface VisitOrDriver {
 
     /**
      * Set the next PlanningVisit to this planning entity.
+     *
      * @param nextPlanningVisit The next planning visit (CustomerVisit, RestaurantVisit instance)
      */
     void setNextPlanningVisit(PlanningVisit nextPlanningVisit);
