+++ prototype/dispatch/order-dispatcher/src/main/java/com/aws/proto/dispatching/domain/planningentity/instant/sequential/DeliveryOrDriver.java
@@ -10,10 +10,10 @@
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,22 +23,35 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.aws.proto.dispatching.domain.planningentity.v1;
+package com.aws.proto.dispatching.domain.planningentity.instant.sequential;
 
-import com.aws.proto.dispatching.domain.planningentity.base.Order;
-import com.aws.proto.dispatching.domain.location.CustomerLocation;
+import com.aws.proto.dispatching.util.Constants;
+import org.optaplanner.core.api.domain.entity.PlanningEntity;
+import org.optaplanner.core.api.domain.variable.InverseRelationShadowVariable;
 
-public class CustomerVisit extends PlanningVisit {
-    public CustomerVisit() {
+@PlanningEntity
+public interface DeliveryOrDriver {
+    /**
+     * Get the driver assigned to this planning entity.
+     *
+     * @return The PlanningDriver. Sometimes {@code null}.
+     */
+    PlanningDriver getPlanningDriver();
 
-    }
-    public CustomerVisit(Order order, CustomerLocation location) {
-        super(order, location);
-    }
+    /**
+     * Get the next delivery assigned to this planning entity
+     *
+     * @return The next PlanningDelivery. Sometimes {@code null}.
+     */
+    @InverseRelationShadowVariable(sourceVariableName = Constants.PreviousDeliveryOrDriver)
+    PlanningDelivery getNextPlanningDelivery();
 
-    @Override
-    public String toString() {
-        return String.format("[DROP][order=%8s]",
-          this.getOrder().getShortId());
-    }
+    /**
+     * Set the next PlanningDelivery to this planning entity.
+     *
+     * @param nextPlanningDelivery The next planning delivery (CustomerVisit, RestaurantVisit instance)
+     */
+    void setNextPlanningDelivery(PlanningDelivery nextPlanningDelivery);
+
+    boolean isDriver();
 }
