+++ prototype/dispatch/order-dispatcher/src/main/java/com/aws/proto/dispatching/planner/solution/instant/sequential/DispatchingConstraintProvider.java
@@ -10,10 +10,10 @@
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,10 +23,10 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.aws.proto.dispatching.planner.solution.v2;
+package com.aws.proto.dispatching.planner.solution.instant.sequential;
 
-import com.aws.proto.dispatching.domain.planningentity.v2.PlanningDelivery;
-import com.aws.proto.dispatching.domain.planningentity.v2.PlanningDriver;
+import com.aws.proto.dispatching.domain.planningentity.instant.sequential.PlanningDelivery;
+import com.aws.proto.dispatching.domain.planningentity.instant.sequential.PlanningDriver;
 import org.optaplanner.core.api.score.buildin.hardmediumsoftlong.HardMediumSoftLongScore;
 import org.optaplanner.core.api.score.stream.Constraint;
 import org.optaplanner.core.api.score.stream.ConstraintFactory;
@@ -40,8 +40,8 @@ public class DispatchingConstraintProvider implements ConstraintProvider {
     public Constraint[] defineConstraints(ConstraintFactory constraintFactory) {
         return new Constraint[]{
 //          olderOrdersCannotBeAssignedAfterFresher(constraintFactory),
-          dontAssignMultipleIfOtherDriversUnassigned(constraintFactory),
-          preferCloserDriverToPickupLocation(constraintFactory)
+                dontAssignMultipleIfOtherDriversUnassigned(constraintFactory),
+                preferCloserDriverToPickupLocation(constraintFactory)
 //          preferOccupiedDrivers(constraintFactory)
         };
     }
@@ -52,37 +52,37 @@ public class DispatchingConstraintProvider implements ConstraintProvider {
 
     protected Constraint olderOrdersCannotBeAssignedAfterFresher(ConstraintFactory factory) {
         return factory.from(PlanningDelivery.class)
-          .join(PlanningDelivery.class,
-                Joiners.equal(PlanningDelivery::getPlanningDriver, PlanningDelivery::getPlanningDriver),
-                Joiners.equal(PlanningDelivery::getNextPlanningDelivery, Function.identity()),
-                Joiners.greaterThan(PlanningDelivery::getOrderTimestamp, PlanningDelivery::getOrderTimestamp),
-                Joiners.filtering((a, b) -> a.getPlanningId() != b.getPlanningId())
-            )
-          .penalize(
-            "olderOrdersCannotBeAssignedAfterFresher",
-            HardMediumSoftLongScore.ONE_HARD
-          );
+                .join(PlanningDelivery.class,
+                        Joiners.equal(PlanningDelivery::getPlanningDriver, PlanningDelivery::getPlanningDriver),
+                        Joiners.equal(PlanningDelivery::getNextPlanningDelivery, Function.identity()),
+                        Joiners.greaterThan(PlanningDelivery::getOrderTimestamp, PlanningDelivery::getOrderTimestamp),
+                        Joiners.filtering((a, b) -> a.getPlanningId() != b.getPlanningId())
+                )
+                .penalize(
+                        "olderOrdersCannotBeAssignedAfterFresher",
+                        HardMediumSoftLongScore.ONE_HARD
+                );
     }
 
     protected Constraint dontAssignMultipleIfOtherDriversUnassigned(ConstraintFactory factory) {
         return factory.from(PlanningDriver.class)
-          .join(PlanningDriver.class,
-            Joiners.filtering((a, b) -> a.chainLength() == 0 && b.chainLength() > 1),
-            Joiners.filtering((a, b) -> a.getId() != b.getId())
-            )
-          .penalize(
-            "dontAssignMultipleIfOtherDriversUnassigned",
-            HardMediumSoftLongScore.ONE_HARD
-          );
+                .join(PlanningDriver.class,
+                        Joiners.filtering((a, b) -> a.chainLength() == 0 && b.chainLength() > 1),
+                        Joiners.filtering((a, b) -> a.getId() != b.getId())
+                )
+                .penalize(
+                        "dontAssignMultipleIfOtherDriversUnassigned",
+                        HardMediumSoftLongScore.ONE_HARD
+                );
     }
 
     protected Constraint deliveryPickupMustBeAfterPreviousDropoffPlusTravel(ConstraintFactory factory) {
         return factory.from(PlanningDelivery.class)
-          .penalizeLong(
-            "deliveryPickupMustBeAfterPreviousDropoffPlusTravel",
-            HardMediumSoftLongScore.ONE_HARD,
-            PlanningDelivery::lateFromFromPreviousDelivery
-          );
+                .penalizeLong(
+                        "deliveryPickupMustBeAfterPreviousDropoffPlusTravel",
+                        HardMediumSoftLongScore.ONE_HARD,
+                        PlanningDelivery::lateFromFromPreviousDelivery
+                );
     }
 
 //    protected Constraint prioritizeDriversWithHigherXScore(ConstraintFactory factory) {
@@ -109,16 +109,17 @@ public class DispatchingConstraintProvider implements ConstraintProvider {
 
     /**
      * OBSOLETE
+     *
      * @param factory
      * @return
      */
     protected Constraint preferOccupiedDrivers(ConstraintFactory factory) {
         return factory.from(PlanningDriver.class)
-          .penalizeLong(
-            "preferOccupiedDrivers",
-            HardMediumSoftLongScore.ONE_MEDIUM,
-            PlanningDriver::scoreForOccupancy
-          );
+                .penalizeLong(
+                        "preferOccupiedDrivers",
+                        HardMediumSoftLongScore.ONE_MEDIUM,
+                        PlanningDriver::scoreForOccupancy
+                );
     }
 
     // ************************************************************************
@@ -137,14 +138,14 @@ public class DispatchingConstraintProvider implements ConstraintProvider {
 //          );
 
         return factory.from(PlanningDelivery.class)
-          .join(PlanningDriver.class,
-            Joiners.equal(PlanningDelivery::getPlanningDriver, PlanningDriver::getPlanningDriver)
+                .join(PlanningDriver.class,
+                        Joiners.equal(PlanningDelivery::getPlanningDriver, PlanningDriver::getPlanningDriver)
 //            Joiners.equal(PlanningDelivery::self, PlanningDriver::getNextPlanningDelivery)
-          )
-          .penalizeLong(
-            "preferCloserDriverToPickupLocation",
-            HardMediumSoftLongScore.ONE_MEDIUM,
-            (delivery, driver) -> delivery.scoreForDistance(driver.getLocation())
-          );
+                )
+                .penalizeLong(
+                        "preferCloserDriverToPickupLocation",
+                        HardMediumSoftLongScore.ONE_MEDIUM,
+                        (delivery, driver) -> delivery.scoreForDistance(driver.getLocation())
+                );
     }
 }
