+++ prototype/dispatch/order-dispatcher/src/main/java/com/aws/proto/dispatching/planner/solution/instant/sequential/DispatchingSolution.java
@@ -10,10 +10,10 @@
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,10 +23,10 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.aws.proto.dispatching.planner.solution.v1;
+package com.aws.proto.dispatching.planner.solution.instant.sequential;
 
-import com.aws.proto.dispatching.domain.planningentity.v1.PlanningDriver;
-import com.aws.proto.dispatching.domain.planningentity.v1.PlanningVisit;
+import com.aws.proto.dispatching.domain.planningentity.instant.sequential.PlanningDelivery;
+import com.aws.proto.dispatching.domain.planningentity.instant.sequential.PlanningDriver;
 import com.aws.proto.dispatching.util.Constants;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.optaplanner.core.api.domain.solution.PlanningEntityCollectionProperty;
@@ -47,21 +47,27 @@ public class DispatchingSolution {
 
     private String name;
 
+    private Long createdAt;
+    private String executionId;
+
     private List<PlanningDriver> planningDrivers;
-    private List<PlanningVisit> planningVisits;
+    private List<PlanningDelivery> planningDeliveries;
 
     private HardMediumSoftLongScore score;
 
-    public DispatchingSolution() {}
-
-    public DispatchingSolution(UUID id, String name, List<PlanningDriver> planningDrivers, List<PlanningVisit> planningVisits) {
-        this.id = id;
-        this.name = name;
-        this.planningDrivers = planningDrivers;
-        this.planningVisits = planningVisits;
+    public DispatchingSolution() {
     }
 
-    //    @PlanningEntityCollectionProperty
+    public DispatchingSolution(UUID id, String name, long createdAt, String executionId, List<PlanningDriver> planningDrivers, List<PlanningDelivery> planningDeliveries) {
+        this.id = id;
+        this.name = name;
+        this.createdAt = createdAt;
+        this.executionId = executionId;
+        this.planningDrivers = planningDrivers;
+        this.planningDeliveries = planningDeliveries;
+    }
+
+//    @PlanningEntityCollectionProperty
 //    protected List<DeliveryPlan> deliveryPlans;
 //    @ProblemFactCollectionProperty
 //    protected List<MultiPDDelivery> multiPDDeliveries;
@@ -82,6 +88,14 @@ public class DispatchingSolution {
         this.name = name;
     }
 
+    public Long getCreatedAt() {
+        return createdAt;
+    }
+
+    public String getExecutionId() {
+        return executionId;
+    }
+
     @ProblemFactCollectionProperty
     @ValueRangeProvider(id = Constants.PlanningDriverRange)
     public List<PlanningDriver> getPlanningDrivers() {
@@ -93,13 +107,13 @@ public class DispatchingSolution {
     }
 
     @PlanningEntityCollectionProperty
-    @ValueRangeProvider(id = Constants.PlanningVisitRange)
-    public List<PlanningVisit> getPlanningVisits() {
-        return planningVisits;
+    @ValueRangeProvider(id = Constants.PlanningDeliveryRange)
+    public List<PlanningDelivery> getPlanningDeliveries() {
+        return planningDeliveries;
     }
 
-    public void setPlanningVisits(List<PlanningVisit> planningVisits) {
-        this.planningVisits = planningVisits;
+    public void setPlanningDeliveries(List<PlanningDelivery> planningDeliveries) {
+        this.planningDeliveries = planningDeliveries;
     }
 
     @PlanningScore
