+++ prototype/dispatch/order-dispatcher/src/main/resources/application-docker.properties
@@ -42,13 +42,11 @@ app.client.driver-api.apikey-secret-name=GeoTrackingApiKeySecret
 app.driver-query.initial-radius-in-m=500
 app.driver-query.extend-radius-in-m=500
 app.driver-query.max-request-count=10
-
 org.jobrunr.background-job-server.enabled=false
 org.jobrunr.background-job-server.worker_count=24 #this value normally is defined by the amount of CPU's that are available
 org.jobrunr.background-job-server.poll_interval=15 #check for new work every 15 seconds
 org.jobrunr.background-job-server.delete_succeeded_jobs_after=36 #succeeded jobs will go to the deleted state after 36 hours
 org.jobrunr.background-job-server.permanently_delete_deleted_jobs_after=72 #deleted jobs will be deleted permanently after 72 hours
-
 ## SOLVER - rest
 quarkus.http.port=8080
 ## :: CORS
@@ -84,14 +82,13 @@ quarkus.log.file.rotation.rotate-on-boot=true
 #
 # quarkus.optaplanner.solver.termination.spent-limit=1m
 #quarkus.optaplanner.solver.termination.best-score-limit=0hard/*soft
-quarkus.optaplanner.solver-config-xml=dispatchSolverConfig-v2.xml
-
+quarkus.optaplanner.solver-config-xml=instant/sequential/dispatchSolverConfig.xml
 ## DDB config
 quarkus.dynamodb.aws.region=ap-southeast-1
 quarkus.dynamodb.aws.credentials.type=default
 app.ddb.table.assignments=devproto-dispatcher-assignments
 app.ddb.table.dem-area-dispatcher-settings=devproto-demographic-area-dispatch-settings
-  # devproto-demographic-area-dispatch-settings
+# devproto-demographic-area-dispatch-settings
 #
 ## DEV PROFILE
 #
@@ -101,4 +98,4 @@ app.ddb.table.dem-area-dispatcher-settings=devproto-demographic-area-dispatch-se
 %dev.aws.region=ap-southeast-1
 %dev.app.routing.s3-bucket-name=devproto-dispatch-engine-XXXXXXXXXXXX-ap-southeast-1
 %dev.quarkus.http.port=8888
-%dev.quarkus.optaplanner.solver-config-xml=dispatchSolverConfig-v2.xml
+%dev.quarkus.optaplanner.solver-config-xml=instant/sequential/dispatchSolverConfig.xml
