+++ prototype/dispatch/order-dispatcher/src/main/resources/application.properties
@@ -33,7 +33,6 @@ app.routing.local-osm-dir=/home/ec2-user/proto/.graphhopper/openstreetmap
 app.routing.osm-file=indonesia-latest.osm.pbf
 app.routing.s3-osm-keypath=maps/indonesia-latest.osm.pbf
 app.routing.local-graphhopper-dir=/home/ec2-user/.graphhopper/graphhopper/indonesia
-
 app.routing.s3-bucket-name=devproto-dispatch-engine-XXXXXXXXXXXX-ap-southeast-1
 ####
 ## CLIENT PROPS
@@ -43,7 +42,6 @@ app.client.driver-api.apikey-secret-name=GeoTrackingApiKeySecret
 app.driver-query.initial-radius-in-m=500
 app.driver-query.extend-radius-in-m=500
 app.driver-query.max-request-count=10
-
 quarkus.vertx.max-worker-execute-time=360
 org.jobrunr.background-job-server.enabled=false
 org.jobrunr.background-job-server.worker_count=24 #this value normally is defined by the amount of CPU's that are available
@@ -86,15 +84,13 @@ quarkus.log.file.rotation.rotate-on-boot=true
 #
 # quarkus.optaplanner.solver.termination.spent-limit=1m
 #quarkus.optaplanner.solver.termination.best-score-limit=0hard/*soft
-quarkus.optaplanner.solver-config-xml=dispatchSolverConfig-v2.xml
-#quarkus.optaplanner.solver-config-xml=dispatchSolverConfig-v1.xml
-
+quarkus.optaplanner.solver-config-xml=instant/sequential/dispatchSolverConfig.xml
+#quarkus.optaplanner.solver-config-xml=instant/mixedpickupdropoff/dispatchSolverConfig.xml
 ## DDB config
 quarkus.dynamodb.aws.region=ap-southeast-1
 quarkus.dynamodb.aws.credentials.type=default
 app.ddb.table.assignments=devproto-dispatcher-assignments
 app.ddb.table.dem-area-dispatcher-settings=devproto-demographic-area-dispatch-settings
-
 #
 ## DEV PROFILE
 #
@@ -104,5 +100,5 @@ app.ddb.table.dem-area-dispatcher-settings=devproto-demographic-area-dispatch-se
 %dev.aws.region=ap-southeast-1
 %dev.app.routing.s3-bucket-name=devproto-dispatch-engine-XXXXXXXXXXXX-ap-southeast-1
 %dev.quarkus.http.port=8888
-%dev.quarkus.optaplanner.solver-config-xml=dispatchSolverConfig-v2.xml
-#%dev.quarkus.optaplanner.solver-config-xml=dispatchSolverConfig-v1.xml
+%dev.quarkus.optaplanner.solver-config-xml=instant/sequential/dispatchSolverConfig.xml
+#%dev.quarkus.optaplanner.solver-config-xml=instant/mixedpickupdropoff/dispatchSolverConfig.xml
